// src/types/era/IEraPerson.ts
export interface IEraPerson {
    birthday: number;
    creator: string;
    gender: number;
    race: string;
    skinColor: string;
    item_type: string;
    hairColor: string;
    description: string;
    birthLongitude: number;
    type1: number;
    deathday: number;
    type0: number;
    reference: string;
    item_type_sub: string;
    eyeColor: string;
    name: string;
    isConfirmed: boolean;
    birthLatitude: number;
    key: number;
    timestamp: number;
    height: number;
}